Implemented plugin for power governing decider.

+++ src/RAPLPlatform.cpp
@@ -150,7 +150,7 @@ namespace geopm
 
         if((m_control_domain_type != GEOPM_CONTROL_DOMAIN_POWER) &&
            (m_imp->power_control_domain() == GEOPM_DOMAIN_PACKAGE) &&
-           (m_num_package != target.size())) {
+           (m_num_package != (int)target.size())) {
             control_type = GEOPM_TELEMETRY_TYPE_PKG_ENERGY;
             for (int i = 0; i < m_num_package; ++i) {
                 m_imp->write_control(m_imp->power_control_domain(), i, control_type, target[i]);
@@ -163,7 +163,7 @@ namespace geopm
             if (m_imp->power_control_domain() != GEOPM_DOMAIN_PACKAGE) {
                 throw geopm::Exception("RAPLPlatform::enforce_policy: RAPLPlatform Currently only supports package level power", GEOPM_ERROR_NOT_IMPLEMENTED, __FILE__, __LINE__);
             }
-            if (m_num_package != target.size()) {
+            if (m_num_package != (int)target.size()) {
                 throw geopm::Exception("RAPLPlatform::enforce_policy: Policy size does not match domains of control", GEOPM_ERROR_INVALID, __FILE__, __LINE__);
             }
         }
